* commit '510616d215372e59e5e8c39b8de1125fc5b54c60':
  AArch64: Use long for pointers in VideoEditor classes

+++ media/java/android/media/videoeditor/MediaArtistNativeHelper.java
@@ -117,7 +117,7 @@ class MediaArtistNativeHelper {
     private boolean mErrorFlagSet = false;
 
     @SuppressWarnings("unused")
-    private int mManualEditContext;
+    private long mManualEditContext;
 
     /* Listeners */
